WIP: noderesourcetopology: add cache and reserve
Lay down the groundwork to introduce the proper cache.
Add the reserve plugin and introduce all the hook points.
Unfortunately there is no way atm to avoid to have the interlink
between the configuration option and the reserve plugin:
both need to be enabled, and failing to provide a inconsistent
configuration would lead to a not working configuration.

Signed-off-by: Francesco Romani <fromani@redhat.com>